Remove all the Hanami::Action modules in favor of one main module
