+++ lib/hanami/action/cache/expires.rb
@@ -50,8 +50,8 @@ module Hanami
         #
         # @see Hanami::Action#finish
         def finish
-          super
           headers.merge!(self.class.expires_directives.headers) unless headers.include? HEADER
+          super
         end
 
         # Class which stores Expires directives
